Move scripts to hardhat tasks
